refactor: extract cron auth into shared verifyCronAuth utility
