+++ src/app/api/cron/cleanup-active-players/route.ts
@@ -1,26 +1,10 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getSupabaseAdmin } from "@/lib/supabase";
-import crypto from "crypto";
-
-function timingSafeEqual(a: string, b: string): boolean {
-  const bufA = Buffer.from(a);
-  const bufB = Buffer.from(b);
-  if (bufA.byteLength !== bufB.byteLength) return false;
-  return crypto.timingSafeEqual(bufA, bufB);
-}
+import { verifyCronAuth } from "@/lib/cron-auth";
 
 export async function GET(request: NextRequest) {
-  const secret = process.env.CRON_SECRET;
-  if (!secret) {
-    return NextResponse.json({ error: "Server misconfigured" }, { status: 500 });
-  }
-
-  const auth = request.headers.get("authorization") ?? "";
-  const expected = `Bearer ${secret}`;
-
-  if (auth.length !== expected.length || !timingSafeEqual(auth, expected)) {
-    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-  }
+  const authError = verifyCronAuth(request);
+  if (authError) return authError;
 
   const sb = getSupabaseAdmin();
   const cutoff = new Date(Date.now() - 60 * 1000).toISOString();
